refactor(cli): flatten mail command module
Move the Mail command slice to the neutral command root and update its internal export without changing public CLI or API behavior.
