+++ turbo/apps/cli/src/commands/mail/index.ts
@@ -4,7 +4,7 @@ import { connectCommand } from "./connect";
 import { linkCommand } from "./link";
 import { listCommand } from "./list";
 
-export const zeroMailCommand = new Command()
+export const mailCommand = new Command()
   .name("mail")
   .description("Review linked Gmail drafts")
   .addCommand(listCommand)
