+++ turbo/apps/cli/src/commands/mail/shared.ts
@@ -1,5 +1,5 @@
 import type { ZeroMailProvider } from "@okouai/api-contracts/contracts/zero-mail";
-import { getOkouAgentId } from "../../../lib/okou-env";
+import { getOkouAgentId } from "../../lib/okou-env";
 
 export const MAIL_CONNECTOR_SLUG_BY_PROVIDER = {
   gmail: "gmail",
